+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # AEMO Gem Changelog
 
+## 0.2.0
+
+* Default NEM12 parsing to use strict validation
+* Only enforce validation of NMIConfiguration in strict mode
+
 ## 0.1.45
 
 * [Nokogiri vulnerability](https://rubysec.com/advisories/nokogiri-CVE-2017-15412) update